refactor: reorganize magnifier event handling
