+++ src/composables/MultiPageViewerEnhancer/ImageMagnifier/useMagnifierPan.ts
